Moved ServerAssets to tests
