+++ algebras/algebra/src/test/scala/endpoints4s/algebra/AssetsTestApi.scala
@@ -1,5 +1,6 @@
 package endpoints4s.algebra
 
+import server.ServerAssets
 trait AssetsTestApi extends EndpointsTestApi with ServerAssets {
 
   val assetEndpoint: Endpoint[AssetRequest, AssetResponse] =
