+++ algebras/algebra/src/test/scala/endpoints4s/algebra/server/ServerAssetTest.scala
@@ -2,7 +2,7 @@ package endpoints4s.algebra.server
 
 import endpoints4s.algebra
 
-trait ServerAssetTest[T <: algebra.Endpoints with algebra.ServerAssets] extends ServerTestBase[T] {
+trait ServerAssetTest[T <: algebra.Endpoints with ServerAssets] extends ServerTestBase[T] {
 
   def serveAssetsEndpoint(
       endpoint: serverApi.Endpoint[serverApi.AssetRequest, serverApi.AssetResponse],
